+++ module/Formation/src/Formation/Entity/Db/FormationInstanceFrais.php
@@ -2,12 +2,15 @@
 
 namespace Formation\Entity\Db;
 
+use Application\Entity\Db\Interfaces\HasSourceInterface;
+use Application\Entity\Db\Traits\HasSourceTrait;
 use UnicaenUtilisateur\Entity\HistoriqueAwareInterface;
 use UnicaenUtilisateur\Entity\HistoriqueAwareTrait;
 
-class FormationInstanceFrais implements HistoriqueAwareInterface
+class FormationInstanceFrais implements HistoriqueAwareInterface, HasSourceInterface
 {
     use HistoriqueAwareTrait;
+    use HasSourceTrait;
 
     /** @var integer */
     private $id;